java: Move RpcClient test suite to flavor-agnostic location.
